Request airdrop inside fullnode and not the script
