+++ src/lib.rs
@@ -14,6 +14,7 @@ pub mod banking_stage;
 pub mod blob_fetch_stage;
 pub mod budget;
 pub mod choose_gossip_peer_strategy;
+pub mod client;
 pub mod crdt;
 pub mod drone;
 pub mod entry;
@@ -51,6 +52,7 @@ pub mod transaction;
 pub mod tvu;
 pub mod vote_stage;
 pub mod voting;
+pub mod wallet;
 pub mod window_stage;
 pub mod write_stage;
 extern crate bincode;
